UPDATE is working: CRUD fully fonctionnal.

+++ admin/controleur/a-delete_bat.php
@@ -1,13 +1,15 @@
 <?php
+if (Passerelle::logged()){
+	if(!empty($_GET["id"])){
+		$id=htmlspecialchars($_GET["id"]);
+		$bat= new Bateau($id, null, null);
+		$bat->retrieve();
+		$bat->delete();
+		$resultat = true;
+	}else{
+		$resultat = false;
+	}
+	$etat = "delete_ok";
 
-if(!empty($_GET["id"])){
-	$id=htmlspecialchars($_GET["id"]);
-	$bat= new Bateau($id, null, null);
-	$bat->retrieve();
-	$bat->delete();
-	$resultat = true;
-}else{
-	$resultat = false;
-}
-$etat = "delete_ok";
- ?>
+}else{ $etat="404"; }
+?>
